+++ indigo_za/static/javascript/indigo/indiga_za/traditions.js
@@ -1,16 +1,8 @@
 (function(exports) {
   "use strict";
 
-  // South Africa
+  // South Africa, relies on the defaults
   exports.Indigo.traditions.za = new Indigo.Tradition({
     country: 'za',
-    grammar: Indigo.traditions.default.settings.grammar,
-    annotatable: Indigo.traditions.default.settings.annotatable,
-    toc: {
-      elements: Indigo.traditions.default.settings.toc.elements,
-      titles: {
-        // just rely on the defaults
-      },
-    },
   });
 })(window);